refactor(agents): add AgentMcpDefinitionService and remove SyncAgentMcpServers (NV-8053)
Centralize agent-definition MCP projection and reconcile in a dedicated service,
and delete the sync use case now that enable/disable paths call reconcile directly.
