+++ apps/api/src/app/agents/management/usecases/update-agent-runtime-config/update-agent-runtime-config.usecase.ts
@@ -2,8 +2,8 @@ import { Injectable, NotFoundException, UnprocessableEntityException } from '@ne
 import { FeatureFlagsService, PinoLogger, resolveAgentRuntime } from '@novu/application-generic';
 import { AgentMcpServerRepository, AgentRepository, IntegrationRepository } from '@novu/dal';
 import { AGENT_RUNTIME_PROVIDERS } from '@novu/shared';
-import { belongsOnAgentDefinition, projectMcpRowsToCatalog } from '../../../mcp/project-mcp-servers';
 import { resolveManagedAgentAlwaysAllowToolPermissions } from '../../../mcp/resolve-managed-agent-always-allow-tool-permissions';
+import { AgentMcpDefinitionService } from '../../../mcp/runtime/agent-mcp-definition.service';
 import type {
   AgentRuntimeCapabilitiesDto,
   AgentRuntimeConfigResponseDto,
@@ -17,6 +17,7 @@ export class UpdateAgentRuntimeConfig {
     private readonly integrationRepository: IntegrationRepository,
     private readonly agentMcpServerRepository: AgentMcpServerRepository,
     private readonly featureFlagsService: FeatureFlagsService,
+    private readonly agentMcpDefinitionService: AgentMcpDefinitionService,
     private readonly logger: PinoLogger
   ) {
     this.logger.setContext(this.constructor.name);
@@ -78,6 +79,7 @@ export class UpdateAgentRuntimeConfig {
       useAlwaysAllowToolPermissions,
     });
 
+    // Same as get runtime config: shared agent MCPs only.
     const mcpRows = await this.agentMcpServerRepository.findByAgent({
       organizationId: command.organizationId,
       environmentId: command.environmentId,
@@ -85,9 +87,9 @@ export class UpdateAgentRuntimeConfig {
       enabledOnly: true,
     });
 
-    const mcpServers = projectMcpRowsToCatalog(mcpRows.filter(belongsOnAgentDefinition), this.logger, {
+    const mcpServers = this.agentMcpDefinitionService.project(mcpRows, {
       agentId: agent._id,
-      useCase: UpdateAgentRuntimeConfig.name,
+      caller: UpdateAgentRuntimeConfig.name,
     });
 
     const providerEntry = AGENT_RUNTIME_PROVIDERS.find((p) => p.providerId === providerId);
