+++ apps/api/src/app/agents/mcp/index.ts
@@ -1,3 +1,4 @@
 export { McpConnectionVaultService } from './connections/mcp-connection-vault.service';
 export { AgentsMcpOAuthController } from './oauth/agents-mcp-oauth.controller';
 export { McpOAuthDiscoveryService } from './oauth/mcp-oauth-discovery.service';
+export { AgentMcpDefinitionService } from './runtime/agent-mcp-definition.service';
