--- apps/api/src/app/agents/mcp/servers/sync-agent-mcp-servers/sync-agent-mcp-servers.command.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-import { EnvironmentCommand } from '../../../../shared/commands/project.command';
-
-export class SyncAgentMcpServersCommand extends EnvironmentCommand {
-  @IsNotEmpty()
-  @IsString()
-  agentId: string;
-}
